move pomodoros progress bar reset button to the component
