+++ src/components/PomodorosBar.vue
@@ -1,5 +1,13 @@
 <template>
-  <div class="pomodoros-progress-bar" :style="progressBarHeight"></div>
+  <div class="pomodoros-progress-bar" :style="progressBarHeight">
+    <button
+      v-if="pomodorosTotal"
+      class="pomodoros-progress-bar__reset-button btn btn-link"
+      @click="setPomodorosTotal(0)"
+    >
+      Reset
+    </button>
+  </div>
 </template>
 
 <script>
@@ -37,4 +45,15 @@ export default {
   width: 100vw
   transition: 0.3s
   background-color: var(--light)
+
+  .pomodoros-progress-bar__reset-button
+    position: absolute
+    bottom: 0
+    right: 0
+    z-index: 1000
+    font-size: 4vh
+    color: var(--dark)
+    &:hover
+      color: var(--super-dark)
+
 </style>
